Added java doc to DAO and model classes. Moved some into outOfScope // Session is now in model/helper (fits better with last changes)
