+++ src/main/java/de/badwalden/schule/dao/DBConnector.java
@@ -1,38 +1,56 @@
 package de.badwalden.schule.dao;
 
-import de.badwalden.schule.model.Student;
 import de.badwalden.schule.ui.helper.DialogHelper;
 import de.badwalden.schule.ui.helper.LanguageHelper;
 import io.github.cdimascio.dotenv.Dotenv;
-import javafx.animation.KeyFrame;
-import javafx.animation.Timeline;
 import javafx.scene.control.Alert;
-import javafx.util.Duration;
-
 import java.sql.*;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+
+/**
+ * Singleton class to handle database connections and operations.
+ */
 public class DBConnector {
     private static final Logger logger = Logger.getLogger(DBConnector.class.getName());
     private Connection connection;
     private static DBConnector instance;
     private static final Dotenv env = Dotenv.configure().load();
 
+    /**
+     * Private constructor for creating a DBConnector instance.
+     * Initializes the connection by calling the connect method.
+     */
     private DBConnector() {
         this.connection = connect();
     }
 
+    /**
+     * Gets the single instance of the DBConnector class, creating it if it does not yet exist.
+     *
+     * @return the single instance of DBConnector
+     */
     public static synchronized DBConnector getInstance() {
+
         if (instance == null) {
             instance = new DBConnector();
         }
         return instance;
     }
 
+    /**
+     * Establishes a database connection using a connection URL from the environment configuration.
+     * Throws IllegalArgumentException if the URL is missing or empty, and SQLException on connection failure.
+     *
+     * @return a new Connection object
+     * @throws SQLException if a database access error occurs
+     * @throws IllegalArgumentException if the connection URL is null or empty
+     */
     public Connection connect() {
+
         String large_url = env.get("CONNECTION_URL");
 
         try {
@@ -51,7 +69,14 @@ public class DBConnector {
         }
     }
 
+    /**
+     * Attempts to re-establish a database connection if the initial connection fails.
+     * Displays an error dialog if reconnection fails.
+     *
+     * @return a new or null Connection object depending on whether the connection attempt succeeds
+     */
     private Connection scheduleReconnection() {
+
         String large_url = env.get("CONNECTION_URL");
 
         DialogHelper.showTimedAlertDialog(Alert.AlertType.ERROR, LanguageHelper.getString("db_connection_error"),
@@ -61,11 +86,17 @@ public class DBConnector {
             return DriverManager.getConnection(large_url);
         } catch (Exception e) {
             logger.log(Level.SEVERE, LanguageHelper.getString("reconnection_failed"), e);
+            DialogHelper.showAlertDialog(Alert.AlertType.ERROR,"Reconnection failed", "Please Restart the program");
             return null;
         }
     }
 
+    /**
+     * Closes the current database connection.
+     * Logs and shows an error if the closing fails.
+     */
     private void close() {
+
         try {
             this.connection.close();
         } catch (SQLException e) {
@@ -74,7 +105,15 @@ public class DBConnector {
         }
     }
 
+    /**
+     * Executes a SQL query with the specified parameters and returns the result as a list of Object arrays.
+     *
+     * @param sql the SQL query to execute
+     * @param params the parameters for the SQL query
+     * @return a List of Object arrays, each representing a row of the result set
+     */
     public List<Object[]> executeQuery(String sql, Object[] params) {
+
         List<Object[]> results = new ArrayList<>();
 
         if (this.connection != null) {
@@ -100,7 +139,6 @@ public class DBConnector {
         return results;
     }
 
-
     /**
      * Executes an update SQL statement.
      *
@@ -108,6 +146,7 @@ public class DBConnector {
      * @return the number of rows affected by the update
      */
     public int executeUpdate(String sql, Object[] params) {
+
         if (this.connection != null) {
             try (PreparedStatement pstmt = this.connection.prepareStatement(sql)) {
                 for (int i = 0; i < params.length; i++) {
@@ -121,8 +160,6 @@ public class DBConnector {
         }
         return 0;
     }
-
-
 
 }
