+++ src/main/java/de/badwalden/schule/exception/UnexpectedResultsException.java
@@ -1,13 +1,25 @@
 package de.badwalden.schule.exception;
 
+
+/**
+ * Exception thrown when the number of results from a database query does not meet the expected count.
+ */
 public class UnexpectedResultsException extends Exception {
 
     private final int wantedCount;
     private final int realCount;
 
+    /**
+     * Constructs a new exception with the specified detail message, expected count, and actual count.
+     *
+     * @param message the detail message
+     * @param wantedCount the expected number of results
+     * @param realCount the actual number of results obtained
+     */
     public UnexpectedResultsException(String message, int wantedCount, int realCount) {
-        super(message); // Passes the message to the base Exception class
-        this.wantedCount = wantedCount; // Stores the count of results that led to the exception
+
+        super(message);
+        this.wantedCount = wantedCount;
         this.realCount = realCount;
     }
 
@@ -15,5 +27,8 @@ public class UnexpectedResultsException extends Exception {
         return realCount;
     }
 
-    public int getWantedCount() { return wantedCount; }
+    public int getWantedCount() {
+        return wantedCount;
+    }
+
 }
