--- src/main/java/de/badwalden/schule/model/Admin.java
@@ -1,7 +1,0 @@
-package de.badwalden.schule.model;
-
-public class Admin extends User {
-    public Admin() {
-
-    }
-}
