+++ src/main/java/de/badwalden/schule/model/CareOffer.java
@@ -3,17 +3,23 @@ package de.badwalden.schule.model;
 import de.badwalden.schule.dao.CareOfferDAO;
 import de.badwalden.schule.dao.DBConnector;
 import de.badwalden.schule.ui.helper.LanguageHelper;
-import kotlin.collections.ArrayDeque;
-
 import java.util.ArrayList;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+
+/**
+ * Represents a care offer, including details such as supervisory staff and availability.
+ * Implements ModelSyncRequirements for database synchronization.
+ */
 public class CareOffer extends Service implements ModelSyncRequirements {
         private static final Logger logger = Logger.getLogger(DBConnector.class.getName());
         final static CareOfferDAO careOfferDao = new CareOfferDAO();
 
+        /**
+         * Constructs a CareOffer with the specified details.
+         */
         public CareOffer(int id, Supervisor supervisor, int oldestClassLevel, int youngestClassLevel, String careOfferName, String description, int seatsAvailable) {
         super();
 
@@ -27,6 +33,9 @@ public class CareOffer extends Service implements ModelSyncRequirements {
 
         }
 
+        /**
+         * Updates this care offer in the database.
+         */
         @Override
         public void update() {
                 logger.log(Level.INFO, LanguageHelper.getString("update_careoffer"));
@@ -37,7 +46,10 @@ public class CareOffer extends Service implements ModelSyncRequirements {
                 }
         }
 
-        // Method to convert CareOffer properties to an Object array
+        /**
+         * Converts this CareOffer's attributes to an Object array suitable for database operations.
+         * @return an Object array containing care offer details
+         */
         @Override
         public Object[] toObjectArray() {
                 return new Object[]{
