+++ src/main/java/de/badwalden/schule/model/ModelSyncRequirements.java
@@ -1,9 +1,18 @@
 package de.badwalden.schule.model;
 
-import de.badwalden.schule.dao.DatabaseInteractions;
 
+/**
+ * Interface defining requirements for model synchronization.
+ */
 public interface ModelSyncRequirements {
+    /**
+     * Updates the model in the database.
+     */
     public void update();
 
+    /**
+     * Converts the model data into an object array for database operations.
+     * @return an array of objects representing the model data
+     */
     public Object[] toObjectArray();
 }
