+++ src/main/java/de/badwalden/schule/model/outOfScope/Admin.java
@@ -1,0 +1,9 @@
+package de.badwalden.schule.model.outOfScope;
+
+import de.badwalden.schule.model.User;
+
+public class Admin extends User {
+    public Admin() {
+
+    }
+}
