+++ src/main/java/de/badwalden/schule/model/outOfScope/Sclass.java
@@ -1,7 +1,6 @@
 package de.badwalden.schule.model.outOfScope;
 
 import de.badwalden.schule.model.Student;
-import de.badwalden.schule.model.Teacher;
 
 import java.util.List;
