+++ src/main/java/de/badwalden/schule/model/outOfScope/Teacher.java
@@ -1,5 +1,6 @@
-package de.badwalden.schule.model;
+package de.badwalden.schule.model.outOfScope;
 
+import de.badwalden.schule.model.User;
 import de.badwalden.schule.model.outOfScope.Sclass;
 
 import java.util.List;
@@ -8,10 +9,6 @@ public class Teacher extends User {
     private List<Sclass> sclassList;
 
     public Teacher() {
-
-    }
-
-    public void updateSclassImage(String newPath){
 
     }
